use INTEGER*4 in ranlib functions (bug #51961)
Until now we have relied on proper compiler flag settings to compile
these files so that Fortran integer variables are 32-bits wide.
Instead, we now use the non-standard but almost universally available
INTEGER*4 declaration.

* advnst.f, genmn.f, genmul.f, genprm.f, getcgn.f, getsd.f, ignbin.f,
ignlgi.f, ignnbn.f, ignpoi.f, ignuin.f, initgn.f, inrgcm.f, lennob.f,
mltmod.f, phrtsd.f, ranf.f, setall.f, setant.f, setgmn.f, setsd.f,
sexpo.f, snorm.f: Declare all integer variables as INTEGER*4. This
makes the parameters match prototypes in lo-ranlib-proto.h and ensures
that the generators work as intended (no conversions between 8-bit and
4-bit integer values, for example).

* wrap.f: Provide types for external functions.

+++ liboctave/external/ranlib/genmn.f
@@ -39,7 +39,7 @@ C     .. Array Arguments ..
 C     ..
 C     .. Local Scalars ..
       REAL ae
-      INTEGER i,icount,j,p
+      INTEGER*4 i,icount,j,p
 C     ..
 C     .. External Functions ..
       REAL snorm
